Feat: 로그인 api 에러 및 토큰 생성
